requirements.txt
tensorflow-gpu==2.0.0
tensorflow-addons==0.6.0
plotly